Add diffOnly option to approve command (#402)
* Add option "diffOnly" to approve command

* Only copy the changed files instead of moving them

* Document --diffOnly cli flag

* Allow overwriting the option set in config from the cli

* Update default value for cli flag

* Add diffOnly option to configuration

* Optimize selection of images that should be approved

+++ packages/runner/src/commands/approve/index.js
@@ -10,17 +10,40 @@ const isPNG = (file) => file.substr(-4) === '.png';
 
 function approve(args) {
   const config = getConfig();
-  const { outputDir, referenceDir } = parseOptions(args, config);
+  const { outputDir, differenceDir, referenceDir, diffOnly } = parseOptions(
+    args,
+    config
+  );
 
-  const files = fs.readdirSync(outputDir).filter(isPNG);
+  // If diff only is active, only copy over the files that were changed
+  const inputDir = diffOnly ? differenceDir : outputDir;
+  const files = fs.readdirSync(inputDir).filter(isPNG);
+
   if (!files.length) {
     die(
       'No images found to approve',
       'Run update command to generate reference files instead'
     );
   }
+
+  if (diffOnly) {
+    /**
+     * If diff only is active, the reference directory should not be emptied.
+     * Instead only the files that changed will be copied over, overwriting the existing ones.
+     * The files are copied and not moved, so running loki approve without --diffOnly after running it with --diffOnly
+     * would not delete them as they would no longer be present in the current images.
+     */
+    files.forEach((file) =>
+      fs.copySync(path.join(outputDir, file), path.join(referenceDir, file), {
+        overwrite: true,
+      })
+    );
+    return;
+  }
+
   fs.emptyDirSync(referenceDir);
   fs.ensureDirSync(referenceDir);
+
   files.forEach((file) =>
     fs.moveSync(path.join(outputDir, file), path.join(referenceDir, file))
   );
